// src/db/database.types.ts
export type Json = string | number | boolean | null | { [key: string]: Json | undefined } | Json[];

export interface Database {
  graphql_public: {
    Tables: Record<never, never>;
    Views: Record<never, never>;
    Functions: {
      graphql: {
        Args: {
          extensions?: Json;
          operationName?: string;
          query?: string;
          variables?: Json;
        };
        Returns: Json;
      };
    };
    Enums: Record<never, never>;
    CompositeTypes: Record<never, never>;
  };
  public: {
    Tables: {
      ski_spec_notes: {
        Row: {
          content: string;
          created_at: string;
          id: string;
          ski_spec_id: string;
          updated_at: string;
        };
        Insert: {
          content: string;
          created_at?: string;
          id?: string;
          ski_spec_id: string;
          updated_at?: string;
        };
        Update: {
          content?: string;
          created_at?: string;
          id?: string;
          ski_spec_id?: string;
          updated_at?: string;
        };
        Relationships: [
          {
            foreignKeyName: 'ski_spec_notes_ski_spec_id_fkey';
            columns: ['ski_spec_id'];
            isOneToOne: false;
            referencedRelation: 'ski_specs';
            referencedColumns: ['id'];
          },
        ];
      };
      ski_specs: {
        Row: {
          algorithm_version: string;
          created_at: string;
          description: string | null;
          id: string;
          length: number;
          name: string;
          radius: number;
          relative_weight: number;
          surface_area: number;
          tail: number;
          tip: number;
          updated_at: string;
          user_id: string;
          waist: number;
          weight: number;
        };
        Insert: {
          algorithm_version: string;
          created_at?: string;
          description?: string | null;
          id?: string;
          length: number;
          name: string;
          radius: number;
          relative_weight: number;
          surface_area: number;
          tail: number;
          tip: number;
          updated_at?: string;
          user_id: string;
          waist: number;
          weight: number;
        };
        Update: {
          algorithm_version?: string;
          created_at?: string;
          description?: string | null;
          id?: string;
          length?: number;
          name?: string;
          radius?: number;
          relative_weight?: number;
          surface_area?: number;
          tail?: number;
          tip?: number;
          updated_at?: string;
          user_id?: string;
          waist?: number;
          weight?: number;
        };
        Relationships: [];
      };
    };
    Views: Record<never, never>;
    Functions: Record<never, never>;
    Enums: Record<never, never>;
    CompositeTypes: Record<never, never>;
  };
}

type DatabaseWithoutInternals = Omit<Database, '__InternalSupabase'>;

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, 'public'>];
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema['Enums'] | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals;
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions['schema']]['Enums']
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals;
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions['schema']]['Enums'][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema['Enums']
    ? DefaultSchema['Enums'][DefaultSchemaEnumNameOrOptions]
    : never;

export type CompositeTypes<
  PublicCompositeTypeNameOrOptions extends
    | keyof DefaultSchema['CompositeTypes']
    | { schema: keyof DatabaseWithoutInternals },
  CompositeTypeName extends PublicCompositeTypeNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals;
  }
    ? keyof DatabaseWithoutInternals[PublicCompositeTypeNameOrOptions['schema']]['CompositeTypes']
    : never = never,
> = PublicCompositeTypeNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals;
}
  ? DatabaseWithoutInternals[PublicCompositeTypeNameOrOptions['schema']]['CompositeTypes'][CompositeTypeName]
  : PublicCompositeTypeNameOrOptions extends keyof DefaultSchema['CompositeTypes']
    ? DefaultSchema['CompositeTypes'][PublicCompositeTypeNameOrOptions]
    : never;